Project_DiseaseInfoAPI/requirements.txt
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.1
bootstrap4==0.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.40.0
django-bootstrap4==1.0.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-telegrambot==1.0.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
future==0.18.2
gunicorn==20.0.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
python-decouple==3.3
python-telegram-bot==12.2.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.7
whitenoise==4.1.4
wrapt==1.11.2
